requirements.txt
dj-database-url==0.4.2
Django==1.11.7
gunicorn==19.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2017.3
whitenoise==3.3.1
djangorestframework==3.9.1